+++ chain-core/Cargo.toml
@@ -7,11 +7,6 @@ license = "MIT OR Apache-2.0"
 
 [dependencies]
 chain-ser = { path = "../chain-ser" }
-cfg-if = "0.1"
-quickcheck = { version = "0.9", optional = true }
 
 [dev-dependencies]
 quickcheck = "0.9"
-
-[features]
-property-test-api = ["quickcheck"]
