+++ chain-impl-mockchain/Cargo.toml
@@ -34,7 +34,8 @@ with-bench = []
 [dev-dependencies]
 quickcheck = "0.9"
 quickcheck_macros = "0.9"
-chain-core = { path = "../chain-core", features=["property-test-api"]}
+chain-core = { path = "../chain-core"}
+chain-test-utils = { path = "../chain-test-utils" }
 chain-crypto = { path = "../chain-crypto", features=["property-test-api"]}
 chain-time = { path = "../chain-time", features=["property-test-api"]}
 chain-addr = { path = "../chain-addr", features=["property-test-api"]}
