+++ chain-impl-mockchain/src/fragment/test.rs
@@ -32,7 +32,7 @@ fn fragment_serialization_bijection(b: Fragment) -> TestResult {
 
 quickcheck! {
     fn initial_ents_serialization_bijection(config_params: ConfigParams) -> TestResult {
-        property::testing::serialization_bijection_r(config_params)
+        chain_test_utils::property::serialization_bijection_r(config_params)
     }
 }
